Created framework for child clients and implemented them for CustomerIdentity and DocumentLibrary
As discussed in #2, GIDXClient is now broken up into
CustomerIdentityClient and DocumentLibraryClient.  Shared code for
submitting HTTP requests was moved to ClientBase, which all child
clients should inherit.
